move change requests to its own package
