+++ pkg/cli/commands/apps/canvas/update_test.go
@@ -7,13 +7,10 @@ import (
 	"net/http/httptest"
 	"os"
 	"path/filepath"
-	"reflect"
 	"strings"
 	"testing"
 
-	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
-	"github.com/superplanehq/superplane/pkg/cli/core"
 	"github.com/superplanehq/superplane/pkg/openapi_client"
 )
 
@@ -122,71 +119,6 @@ func TestResolveCanvasForFileUpdateWhitespaceIDReturnsError(t *testing.T) {
 	_, _, err := resolveCanvasForFileUpdate(filePath)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "metadata.id is required")
-}
-
-func TestAutoLayoutFlagsWereSet(t *testing.T) {
-	t.Run("nil command", func(t *testing.T) {
-		require.False(t, autoLayoutFlagsWereSet(core.CommandContext{}))
-	})
-	t.Run("auto-layout flag changed", func(t *testing.T) {
-		cmd := cobra.Command{}
-		cmd.Flags().String("auto-layout", "", "")
-		cmd.Flags().String("auto-layout-scope", "", "")
-		cmd.Flags().StringArray("auto-layout-node", nil, "")
-		require.NoError(t, cmd.ParseFlags([]string{"--auto-layout=horizontal"}))
-		require.True(t, autoLayoutFlagsWereSet(core.CommandContext{Cmd: &cmd}))
-	})
-	t.Run("auto-layout-scope flag changed", func(t *testing.T) {
-		cmd := cobra.Command{}
-		cmd.Flags().String("auto-layout", "", "")
-		cmd.Flags().String("auto-layout-scope", "", "")
-		cmd.Flags().StringArray("auto-layout-node", nil, "")
-		require.NoError(t, cmd.ParseFlags([]string{"--auto-layout-scope=full-canvas"}))
-		require.True(t, autoLayoutFlagsWereSet(core.CommandContext{Cmd: &cmd}))
-	})
-	t.Run("auto-layout-node flag changed", func(t *testing.T) {
-		cmd := cobra.Command{}
-		cmd.Flags().String("auto-layout", "", "")
-		cmd.Flags().String("auto-layout-scope", "", "")
-		cmd.Flags().StringArray("auto-layout-node", nil, "")
-		require.NoError(t, cmd.ParseFlags([]string{"--auto-layout-node=a"}))
-		require.True(t, autoLayoutFlagsWereSet(core.CommandContext{Cmd: &cmd}))
-	})
-	t.Run("no auto-layout flags parsed", func(t *testing.T) {
-		cmd := cobra.Command{}
-		cmd.Flags().String("auto-layout", "", "")
-		require.NoError(t, cmd.ParseFlags(nil))
-		require.False(t, autoLayoutFlagsWereSet(core.CommandContext{Cmd: &cmd}))
-	})
-}
-
-func TestParseAutoLayoutRejectsInvalidAlgorithm(t *testing.T) {
-	_, err := parseAutoLayout("vertical", "", nil)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "unsupported auto layout")
-}
-
-func TestParseAutoLayoutRejectsInvalidScope(t *testing.T) {
-	_, err := parseAutoLayout("horizontal", "whole-world", nil)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "unsupported auto layout scope")
-}
-
-func TestParseAutoLayoutScopeFullAlias(t *testing.T) {
-	autoLayout, err := parseAutoLayout("horizontal", "full", nil)
-	require.NoError(t, err)
-	require.NotNil(t, autoLayout)
-	require.Equal(t, openapi_client.CANVASAUTOLAYOUTSCOPE_SCOPE_FULL_CANVAS, autoLayout.GetScope())
-}
-
-func TestParseAutoLayoutDisableAcceptsOffAndNone(t *testing.T) {
-	for _, v := range []string{"off", "none", "NONE"} {
-		t.Run(v, func(t *testing.T) {
-			al, err := parseAutoLayout(v, "", nil)
-			require.NoError(t, err)
-			require.Nil(t, al)
-		})
-	}
 }
 
 func TestUpdateFromFileJSONOutputWhenDraft(t *testing.T) {
@@ -418,44 +350,6 @@ func TestBuildDefaultAutoLayoutUsesFullCanvas(t *testing.T) {
 	}
 	if autoLayout.HasNodeIds() {
 		t.Fatalf("expected no node ids for default full-canvas strategy, got %v", autoLayout.GetNodeIds())
-	}
-}
-
-func TestParseAutoLayoutDefaultsAlgorithmToHorizontal(t *testing.T) {
-	autoLayout, err := parseAutoLayout("", "connected-component", []string{"node-1", " node-2 ", "node-1"})
-	if err != nil {
-		t.Fatalf("parseAutoLayout returned error: %v", err)
-	}
-	if autoLayout == nil {
-		t.Fatalf("expected autoLayout to be set")
-	}
-	if autoLayout.GetAlgorithm() != openapi_client.CANVASAUTOLAYOUTALGORITHM_ALGORITHM_HORIZONTAL {
-		t.Fatalf("expected horizontal auto-layout, got %s", autoLayout.GetAlgorithm())
-	}
-	if autoLayout.GetScope() != openapi_client.CANVASAUTOLAYOUTSCOPE_SCOPE_CONNECTED_COMPONENT {
-		t.Fatalf("expected connected-component scope, got %s", autoLayout.GetScope())
-	}
-	if !reflect.DeepEqual(autoLayout.GetNodeIds(), []string{"node-1", "node-2"}) {
-		t.Fatalf("expected node ids [node-1 node-2], got %v", autoLayout.GetNodeIds())
-	}
-}
-
-func TestParseAutoLayoutDisable(t *testing.T) {
-	autoLayout, err := parseAutoLayout("disable", "", nil)
-	if err != nil {
-		t.Fatalf("parseAutoLayout returned error: %v", err)
-	}
-	if autoLayout != nil {
-		t.Fatalf("expected nil autoLayout when disabled, got %#v", autoLayout)
-	}
-}
-
-func TestParseAutoLayoutDisableRejectsScopeOrNodes(t *testing.T) {
-	if _, err := parseAutoLayout("disable", "connected-component", nil); err == nil {
-		t.Fatalf("expected error when scope is set together with disable")
-	}
-	if _, err := parseAutoLayout("disable", "", []string{"node-1"}); err == nil {
-		t.Fatalf("expected error when node ids are set together with disable")
 	}
 }
