+++ pkg/cli/commands/apps/changes/create.go
@@ -1,0 +1,90 @@
+package changes
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/superplanehq/superplane/pkg/cli/appresolve"
+	"github.com/superplanehq/superplane/pkg/cli/core"
+	"github.com/superplanehq/superplane/pkg/openapi_client"
+)
+
+type CreateCommand struct {
+	versionID   *string
+	title       *string
+	description *string
+}
+
+func (c *CreateCommand) Execute(ctx core.CommandContext) error {
+	if len(ctx.Args) > 1 {
+		return fmt.Errorf("create accepts at most one positional argument")
+	}
+
+	target := ""
+	if len(ctx.Args) == 1 {
+		target = strings.TrimSpace(ctx.Args[0])
+	}
+
+	canvasID, err := resolveCanvasTargetFromOptionalArg(ctx, target)
+	if err != nil {
+		return err
+	}
+
+	versionID := ""
+	if c.versionID != nil {
+		versionID = strings.TrimSpace(*c.versionID)
+	}
+
+	if versionID == "" {
+		changeManagementEnabled, err := appresolve.ChangeManagementEnabled(ctx, canvasID)
+		if err != nil {
+			return err
+		}
+		if !changeManagementEnabled {
+			return fmt.Errorf("change management is disabled for this canvas; enable it in canvas settings to use change requests")
+		}
+
+		versionID, err = appresolve.FindCurrentUserDraftVersionID(ctx, canvasID)
+		if err != nil {
+			return err
+		}
+		if versionID == "" {
+			return fmt.Errorf("no draft version found; run `superplane apps canvas update --draft -f <file>` first")
+		}
+	}
+
+	body := openapi_client.CanvasesCreateCanvasChangeRequestBody{}
+	body.SetVersionId(versionID)
+
+	if c.title != nil {
+		trimmedTitle := strings.TrimSpace(*c.title)
+		if trimmedTitle != "" {
+			body.SetTitle(trimmedTitle)
+		}
+	}
+	if c.description != nil {
+		trimmedDescription := strings.TrimSpace(*c.description)
+		if trimmedDescription != "" {
+			body.SetDescription(trimmedDescription)
+		}
+	}
+
+	response, _, err := ctx.API.CanvasChangeRequestAPI.
+		CanvasesCreateCanvasChangeRequest(ctx.Context, canvasID).
+		Body(body).
+		Execute()
+	if err != nil {
+		return err
+	}
+
+	if response.ChangeRequest == nil {
+		return nil
+	}
+
+	changeRequest := *response.ChangeRequest
+	if !ctx.Renderer.IsText() {
+		return ctx.Renderer.Render(changeRequest)
+	}
+
+	return renderCanvasChangeRequestSummaryText(ctx, "created", changeRequest)
+}
